[Truffle] Fix typo to check for longs when deciding if we can fit a field into a primitive.

+++ core/src/main/java/org/jruby/truffle/runtime/objectstorage/ObjectLayout.java
@@ -81,7 +81,7 @@ public class ObjectLayout {
                 if (type == Integer.class) {
                     canStoreInPrimitive = true;
                     primitivesNeeded = 1;
-                } else if (type == Integer.class) {
+                } else if (type == Long.class) {
                     canStoreInPrimitive = true;
                     primitivesNeeded = 2;
                 } else if (type == Double.class) {
